MOVE GENERATION IS DONE!!! YESSSS
This means move space generation in particular,
meaning it can detect where a piece can move exactly
given it's move range; it does not check for anything
that is more *state* specific, like castling, check,
en passe, etc... YET!

I haven't used automated fuzzing to aggressively to
check thousands of move board states, but I've done manual
testing with every piece, and I'm very confident that there
are no bugs with the exception of rare and obscure cases, that
could never be found by hand.

If it wasn't already obvious, this is a major milestone!